+++ src/modules/voxelrender/tests/VoxelRenderShaderTest.cpp
@@ -5,6 +5,9 @@
 #include "video/tests/AbstractGLTest.h"
 #include "VoxelShader.h"
 #include "VoxelnormShader.h"
+#include "VoxeloitShader.h"
+#include "VoxelnormoitShader.h"
+#include "OitShader.h"
 
 namespace voxelrender {
 
@@ -19,6 +22,24 @@ TEST_P(VoxelRenderShaderTest, testVoxelShader) {
 
 TEST_P(VoxelRenderShaderTest, testVoxelNormShader) {
 	shader::VoxelnormShader shader;
+	EXPECT_TRUE(shader.setup());
+	shader.shutdown();
+}
+
+TEST_P(VoxelRenderShaderTest, testVoxelOitShader) {
+	shader::VoxeloitShader shader;
+	EXPECT_TRUE(shader.setup());
+	shader.shutdown();
+}
+
+TEST_P(VoxelRenderShaderTest, testVoxelNormOitShader) {
+	shader::VoxelnormoitShader shader;
+	EXPECT_TRUE(shader.setup());
+	shader.shutdown();
+}
+
+TEST_P(VoxelRenderShaderTest, testOitShader) {
+	shader::OitShader shader;
 	EXPECT_TRUE(shader.setup());
 	shader.shutdown();
 }